Added presets MenuBarItem
There are some bugs regarding guiupdates i added a method to call but i did not figure out how to call  it from the mandelbrotpanel file.
also the preset panel is now sent to the fractal explorer. long process but this is the best way i could find to do it. could move the logic for the panel to the fractalupdater to require less steps.